+++ internal/bot/workflow.go
@@ -1,22 +1,13 @@
 package bot
 
 import (
-	"context"
-
-	"github.com/outcatcher/scriba/internal/entities"
+	"github.com/outcatcher/scriba/internal/bot/schema"
 	"gopkg.in/telebot.v3"
 )
 
-type UseCases interface {
-	UpdateCountByTelegramID(ctx context.Context, telegramID int64, delta int16) error
-	GetPlayerCountByTelegramID(ctx context.Context, telegramID int64) (int32, error)
-	RegisterWithTelegram(ctx context.Context, telegramID int64) error
-	ListPlayers(ctx context.Context) ([]entities.Player, error)
-}
-
 type workflow interface {
 	// WithUseCases attaches workflow to functionality.
-	WithUseCases(useCases UseCases)
+	WithUseCases(useCases schema.UseCases)
 	// Start is a handler to start workflow.
 	Start(bot *telebot.Bot) telebot.HandlerFunc
 }
